backend/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
eventlet==0.36.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-socketio[client]==5.11.1
